Avoid errors with lengthy captions
Was training on OpenImagesV6 last night and they have a few captions that are more than 256 tokens in length. This (of course) throws an exception which interrupts training and what-not. 

There's probably an argument to be made that you would just want to throw these out altogether, but as it stands I think a trim is a decent enough fix given that's quite a lengthy caption in the first place.