Call LR scheduler explicitly with DeepSpeed
Remove LR scheduler if-guards in training loops (at least in
`train_dalle.py`) in favor of more complicated initialization, fix #278.

The reason we do not modify the `distr_backend.distribute` method is
that we expect its behaviour to be consistent with the backend API.

Scheduling behaviour is now consistent and should not cause issues in
the future if the intervals for scheduling are modified.

What is still confusing: if an LR scheduler is configured in the
DeepSpeed config in `train_dalle.py` **and** `--lr_decay` is passed, the
scheduler will be advanced more often than desired. This could be solved
the same way it's solved in `train_vae.py` if desired.